Decouple the Omnibar from stickiness detection logic
