+++ app/src/main/java/com/duckduckgo/app/browser/commands/Command.kt
@@ -240,4 +240,5 @@ sealed class Command {
     object OpenDuckPlayerSettings : Command()
     object OpenDuckPlayerOverlayInfo : Command()
     object OpenDuckPlayerPageInfo : Command()
+    data object MakeOmnibarStickyIfNeeded : Command()
 }
